Add support for JWKS endpoint
OAuth resource servers implemented with Spring Security 5.x rely on the
JWKS endpoint (/.well-known/jwks.json by default) to obtain JWT keys
required for token verification.

Spring Security OAuth does not support this endpoint out of the box so
we add it explicitly and configure security to allow all clients to
access it.

Note that in production the actual keys shall be obtained from a secured
 storage rather than from the source code as in this example.

Signed-off-by: Oleg Iavorskyi <oleg.iavorskyi@gmail.com>